core/requirements.txt
asgiref==3.9.1
certifi==2025.7.14
charset-normalizer==3.4.2
Django==5.2.4
django-admin-interface==0.30.1
django-ckeditor==6.7.3
django-colorfield==0.14.0
django-filter==25.1
django-jalali-date==2.0.0
django-js-asset==3.1.2
django-mptt==0.17.0
django-phonenumber-field==8.1.0
Faker==37.4.2
idna==3.10
idpay==0.0.1
jalali_core==1.0.0
jdatetime==5.2.0
phonenumbers==9.0.9
pillow==11.3.0
psycopg2-binary==2.9.10
python-decouple==3.8
python-slugify==8.0.4
requests==2.32.4
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
